pyproject.toml
[tool.poetry]
name = "vm_changer"
version = "0.1.0"
description = ""
authors = ["taimaskhanov11 <81189712+taimaskhanov11@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
